pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "browser-agent"
version = "0.1.0"
description = "Autonomous LLM-powered browser agent."
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "Browser Agent Team", email = "team@example.com" }
]
dependencies = [
  "anyio>=4.3",
  "fastapi>=0.115",
  "httpx>=0.27",
  "jinja2>=3.1",
  "orjson>=3.10",
  "playwright>=1.48",
  "pydantic>=2.8",
  "python-dotenv>=1.0",
  "rich>=13.9",
  "typer>=0.12",
  "uvicorn>=0.30",
  "black>=24.10",
  "mypy>=1.11",
  "ruff>=0.6",
  "pytest>=8.3",
  "pytest-asyncio>=0.24"
]

[project.optional-dependencies]
dev = [
  "pytest-mock>=3.14",
  "coverage>=7.6"
]

[project.scripts]
agent-cli = "agent.cli:main"

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
line-length = 95
src = ["src", "tests"]
target-version = "py311"
select = ["E", "F", "I", "B", "UP", "ASYNC", "N"]
ignore = ["B008"]

[tool.mypy]
python_version = "3.11"
mypy_path = "src"
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
no_implicit_optional = true
strict_equality = true
disallow_any_generics = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = [
  "ignore::DeprecationWarning",
  "ignore::pytest.PytestDeprecationWarning"
]
